refactor: generator 패키지명을 strategy로 변경
