+++ src/main/kotlin/lotto/model/lottery/strategy/RandomLotteryNumberGenerator.kt
@@ -1,4 +1,4 @@
-package lotto.model.lottery.generator
+package lotto.model.lottery.strategy
 
 import lotto.model.lottery.LotteryNumber
